project/requirements.txt
# Project dependencies
streamlit
apify-client
pandas
python-dotenv
pillow
numpy
requests